Module hashicorp_vault::client

vault client

Modules

error
Errors

Structs

Auth
Vault auth

PostgresqlLogin
Postgresql secret backend

TokenData
Token data, used in VaultResponse

TokenOptions
Options for creating a token. This is intended to be used as a "builder"-style interface, where you create a new TokenOptions object, call a bunch of chained methods on it, and then pass the result to Client::create_token.

VaultClient
Vault client used to make API requests to the vault

VaultDateTime
Used for responses that return RFC 3339 timestamps See: https://github.com/hashicorp/vault/issues/1654

VaultDuration
Lease duration.

VaultNaiveDateTime
Used for vault responses that return seconds since unix epoch See: https://github.com/hashicorp/vault/issues/1654

VaultResponse
Vault response. Different vault responses have different data types, so D is used to represent this.

WrapData
Wrapped response is serialized json

WrapInfo
Information provided to retrieve a wrapped response

XVaultToken
Token used to authenticate with the vault API

XVaultWrapTTL
The TTL for the token is set by the client using the X-Vault-Wrap-TTL header and can be either an integer number of seconds or a string duration of seconds (15s), minutes (20m), or hours (25h). When using the Vault CLI, you can set this via the -wrap-ttl parameter. Response wrapping is per-request; it is the presence of a value in this header that activates wrapping of the response.