Struct hashicorp_vault::client::TokenOptions

pub struct TokenOptions { /* fields omitted */ }

Options for creating a token. This is intended to be used as a "builder"-style interface, where you create a new TokenOptions object, call a bunch of chained methods on it, and then pass the result to Client::create_token.

use hashicorp_vault::client::{TokenOptions, VaultDuration};

let _ = TokenOptions::default()
  .id("test12345")
  .policies(vec!("root"))
  .default_policy(false)
  .orphan(true)
  .renewable(false)
  .display_name("jdoe-temp")
  .number_of_uses(10)
  .ttl(VaultDuration::hours(3))
  .explicit_max_ttl(VaultDuration::hours(13));Run

If an option is not specified, it will be set accordinding to Vault's standard defaults for newly-created tokens.

Methods

impl TokenOptions

fn id<S: Into<String>>(self, id: S) -> Self

Set the id of the created token to the specified value. This may make it easy for attackers to guess your token. Typically, this is used for testing and similar purposes.

fn policies<'a, I>(self, policies: I) -> Self where I: IntoIterator, I::Item: Into<String>

Supply a list of policies that will be used to grant permissions to the created token. Unless you also call default_policy(false), the policy default will be added to this list in modern versions of vault.

fn default_policy(self, enable: bool) -> Self

Should we grant access to the default policy? Defaults to true.

fn orphan(self, orphan: bool) -> Self

Should this token be an "orphan", allowing it to survive even when the token that created it expires or is revoked?

fn renewable(self, renewable: bool) -> Self

Should the token be renewable?

fn display_name<S>(self, name: S) -> Self where S: Into<String>

For various logging purposes, what should this token be called?

fn number_of_uses(self, uses: u64) -> Self

How many times can this token be used before it stops working?

fn ttl<D: Into<VaultDuration>>(self, ttl: D) -> Self

How long should this token remain valid for?

fn explicit_max_ttl<D: Into<VaultDuration>>(self, ttl: D) -> Self

How long should this token remain valid for, even if it is renewed repeatedly?