Struct hashicorp_vault::client::VaultClient

pub struct VaultClient<'a, T> where T: Decodable {
    pub host: &'a str,
    pub token: String,
    pub data: VaultResponse<T>,
    // some fields omitted
}

Vault client used to make API requests to the vault

Fields
host: &'a str

URL to vault instance

token: String

Token to access vault

data: VaultResponse<T>

Data

Methods

impl<'a> VaultClient<'a, TokenData>

fn new(host: &'a str, token: &'a str) -> Result<VaultClient<'a, TokenData>>

Construct a VaultClient from an existing vault token

impl<'a> VaultClient<'a, ()>

fn new_app_id(host: &'a str, app_id: &'a str, user_id: &'a str) -> Result<VaultClient<'a, ()>>

Construct a VaultClient via the App ID auth backend

impl<'a, T> VaultClient<'a, T> where T: Decodable

fn renew(&mut self) -> Result<()>

Renew lease for VaultClient's token and updates the self.data.auth based upon the response. Corresponds to /auth/token/renew-self.

let host = "http://127.0.0.1:8200";
let token = "test12345";
let mut client = Client::new(host, token).unwrap();

client.renew().unwrap();Run

fn renew_token(&self, token: &str, increment: Option<u64>) -> Result<Auth>

Renew the lease for the specified token. Requires root privileges. Corresponds to /auth/token/renew[/token].

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();

let token_to_renew = "test12345";
client.renew_token(token_to_renew, None).unwrap();Run

fn revoke(self) -> Result<()>

Revoke VaultClient's token. This token can no longer be used. Corresponds to /auth/token/revoke-self.

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();

// Create a temporary token, and use it to create a new client.
let opts = client::TokenOptions::default()
  .ttl(client::VaultDuration::minutes(5));
let res = client.create_token(&opts).unwrap();
let mut new_client = Client::new(host, &res.client_token).unwrap();

// Issue and use a bunch of temporary dynamic credentials.

// Revoke all our dynamic credentials with a single command.
new_client.revoke().unwrap();Run

Note that we consume our self parameter, so you cannot use the client after revoking it.

fn renew_lease(&self, lease_id: &str, increment: Option<u64>) -> Result<VaultResponse<()>>

Renew a specific lease that your token controls. Corresponds to /v1/sys/renew.

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();

// TODO: Right now, we offer no way to get lease information for a
// secret.
let lease_id: String = unimplemented!();

client.renew_lease(&lease_id, None).unwrap();Run

fn lookup(&self) -> Result<VaultResponse<TokenData>>

Lookup token information for this client's token. Corresponds to /auth/token/lookup-self.

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();

let res = client.lookup().unwrap();
assert!(res.data.unwrap().policies.len() >= 0);Run

fn create_token(&self, opts: &TokenOptions) -> Result<Auth>

Create a new vault token using the specified options. Corresponds to /auth/token/create.

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();

let opts = client::TokenOptions::default()
  .display_name("test_token")
  .policies(vec!("root"))
  .default_policy(false)
  .orphan(true)
  .renewable(false)
  .display_name("jdoe-temp")
  .number_of_uses(10)
  .ttl(client::VaultDuration::minutes(1))
  .explicit_max_ttl(client::VaultDuration::minutes(3));
let res = client.create_token(&opts).unwrap();
Run

fn set_secret(&self, key: &str, value: &str) -> Result<()>

Saves a secret

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();
let res = client.set_secret("hello_set", "world");
assert!(res.is_ok());Run

fn get_secret(&self, key: &str) -> Result<String>

Fetches a saved secret

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();
let res = client.set_secret("hello_get", "world");
assert!(res.is_ok());
let res = client.get_secret("hello_get");
assert!(res.is_ok());
assert_eq!(res.unwrap(), "world");Run

fn get_secret_wrapped(&self, key: &str, wrap_ttl: &str) -> Result<VaultResponse<()>>

Fetch a wrapped secret. Token (one-time use) to fetch secret will be in wrap_info.token https://www.vaultproject.io/docs/secrets/cubbyhole/index.html

fn get_cubbyhole_response(&self) -> Result<VaultResponse<HashMap<String, String>>>

Fetch wrapped response from cubbyhole/response

The original response (in the response key) is what is returned

fn delete_secret(&self, key: &str) -> Result<()>

Deletes a saved secret

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();
let res = client.set_secret("hello_delete", "world");
assert!(res.is_ok());
let res = client.delete_secret("hello_delete");
assert!(res.is_ok());Run

fn get_postgresql_backend(&self, name: &str) -> Result<VaultResponse<PostgresqlLogin>>

Get postgresql secret backend https://www.vaultproject.io/docs/secrets/postgresql/index.html

fn policies(&self) -> Result<Vec<String>>

Get a list of policy names defined by this vault. This requires root privileges. Corresponds to /sys/policy.

let host = "http://127.0.0.1:8200";
let token = "test12345";
let client = Client::new(host, token).unwrap();

let res = client.policies().unwrap();
assert!(res.contains(&"root".to_owned()));Run